// src/types/estate.ts
export type Estate = {
  propertyType: string;
  streetAddress: string;
  cityName: string;
  stateOrProvince: string;
  postalCode: string;
  listingPrice: string;
  numberOfBedrooms: string;
  numberOfBathrooms: string;
  generalLocation: string;
  squareFootage: string;
  lotArea: string;
  yearConstructed: string;
  daysListedOnMarket: string;
  pricePerSquareFoot: string;
  mlsIdentifier: string;
  latitude: string;
  longitude: string;
};